rewriting install to zsh

+++ install.zsh
@@ -1,6 +1,12 @@
-#!/bin/sh
+#!/usr/bin/env zsh
 
-help="
+help='
+Usage:
+
+zsh install.zsh [-v|--version] [-h|--help] [-a|--all] \
+  [-b|--bash] [-z|--zsh] [-p|--powerline] [-g|--git] \
+  [-t|--tmux]
+
 Options:
 
 -v/--version          - Print version
@@ -11,15 +17,15 @@ Options:
 -p/--powerline        - Install Powerline config
 -g/--git              - Install Git config
 -t/--tmux             - Install Tmux config
-"
+'
 
-version="
-Version: 0.1.0
-2013, Jose Luis Rivas <me@ghostbar.co>
-Licensed under MIT terms.
-"
+version='
+Version: 0.2.0
+© 2013-2014, Jose Luis Rivas <me@ghostbar.co>
+Licensed under the MIT terms.
+'
 
-bash() {
+function bash {
   [ -e ~/.bashrc ] && mv ~/.bashrc ~/.bashrc.bak
   ln -s $PWD/bashrc ~/.bashrc
 
@@ -36,7 +42,7 @@ bash() {
   ln -s $PWD/aliases ~/.aliases
 }
 
-zsh() {
+function zsh {
   [ -e ~/.zshrc ] && mv ~/.zshrc ~/.zshrc.bak
   ln -s $PWD/zshrc ~/.zshrc
 
@@ -44,13 +50,13 @@ zsh() {
   ln -s $PWD/zsh ~/.zsh
 }
 
-powerline() {
+function powerline {
   [ ! -e ~/.config ] && mkdir ~/.config
   [ -e ~/.config/powerline ] && mv ~/.config/powerline ~/.config/powerline.bak/
   ln -s $PWD/powerline ~/.config/powerline
 }
 
-git() {
+function git {
   [ -e ~/.gitignore ] && mv ~/.gitignore ~/.gitignore.bak
   ln -s $PWD/gitignore ~/.gitignore
 
@@ -58,17 +64,33 @@ git() {
   ln -s $PWD/gitconfig ~/.gitconfig
 }
 
-tmux() {
-  [ -e ~/.tmux.conf ] && mv ~/.tmux.conf ~/.tmux.conf.bak
-  ln -s $PWD/tmux.conf ~/.tmux.conf
+function tmux {
+  print 'Linking tmux configuration';
+
+  if [[ -e ~/.tmux.conf && -e $PWD/tmux.conf ]]; then
+    mv ~/.tmux.conf ~/.tmux.conf.bak;
+  fi
+
+  if [[ -e $PWD/tmux.conf ]]; then
+    ln -s $PWD/tmux.conf ~/.tmux.conf;
+    print "Your tmux configuration is ready to be used."
+  else
+    print "There's no" $PWD/tmux.conf", maybe you are in the wront directory"
+  fi
+
+}
+
+function showHelp {
+  print $version;
+  print $help;
 }
 
 while test -n "$1"; do
   case $1 in
-    -v|--version)
+    (-v|--version)
       echo "$version";
       exit 0;;
-    -a|--all)
+    (-a|--all)
       echo "Install all configs\n";
       bash;
       zsh;
@@ -76,36 +98,34 @@ while test -n "$1"; do
       git;
       tmux;
       exit 0;;
-    -b|--bash)
+    (-b|--bash)
       echo "Install bash config\n";
       bash;
       exit 0;;
-    -z|--zsh)
+    (-z|--zsh)
       echo "Install Zsh config\n";
       bash;
       zsh;
       exit 0;;
-    -p|--powerline)
+    (-p|--powerline)
       echo "Install Powerline config\n";
       powerline;
       exit 0;;
-    -g|--git)
+    (-g|--git)
       echo "Install Git config\n";
       git;
       exit 0;;
-    -t|--tmux)
+    (-t|--tmux)
       echo "Install Tmux config\n";
       tmux;
       exit 0;;
-    *)
-      echo "$help";
-      echo "$version";
+    (*)
+      showHelp;
       exit 0;;
   esac
 done
 
 if test -z $1; then
-  echo "$help"; 
-  echo "$version"; 
+  showHelp;
   exit 0;
 fi
